O programa deletava toda a lista local quando era chamado o alunos.removeif() então substitui por alunos.removeIf(aluno -> ((Aluno)aluno).getMatricula().equals(a.getMatricula()));

+++ src/main/java/com/mycompany/projeto_final/controller/AlunoController.java
@@ -8,8 +8,6 @@ import com.mycompany.projeto_final.exception.AlunoJaCadastradoException;
 import com.mycompany.projeto_final.exception.AlunoNaoEncontradoException;
 import com.mycompany.projeto_final.service.AlunoService;
 import com.mycompany.projeto_final.dao.AlunoDAO;
-import java.sql.SQLIntegrityConstraintViolationException;
-import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
 import javax.swing.table.DefaultTableModel;
@@ -36,7 +34,7 @@ public class AlunoController {
                                     aluno.getCpf(), aluno.getTelefone(), aluno.getIdade());
     }
     
-    public static int quantidadeTotalDeAlunos() {
+    public static long quantidadeTotalDeAlunos() {
         return AlunoService.getSize();
     }
